// src/interfaces/material.interface.ts
export interface Material {
    id: number;
    project_code: string;
    supplier_code: string;
    supplier_name: string;
    material_code: string;
    company: string;
    address: string;
    contact_person: string;
    contact_mobile: string;
    ofc_phone: string;
    description: string;
    quantity: number;
    unit_price: number;
    total_price: number;
    payment_mode: string;
    comment: string;
}